requirements.txt
boto3
python-json-logger
PyYAML
